Tidy and rename helper script
